fix: refactor pluginService
